Created auth and text-editor modules
