--- server/views/index.ejs
@@ -1,4 +1,0 @@
-<h1>Hi <%= name %></h1>
-<form action="/logout?_method=DELETE" method="POST">
-    <button type="submit"> Log Out</button>
-</form>
